requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic-settings>=2.2.1
python-multipart>=0.0.9
pillow>=10.2.0
cachetools>=5.3.3
torch>=2.2.1
torchvision>=0.17.1
opencv-python>=4.9.0.80
requests>=2.31.0
